refactor: outputView.js 파일 만들어 출력 되는 코드 이동
