+++ src/App.js
@@ -3,7 +3,7 @@ import { MainController } from './Controller/mainController.js';
 class App {
   async play() {
     const controller = new MainController();
-    await controller.startProgram();
+    await controller.initializeProgram();
   }
 }
