+++ src/Controller/mainController.js
@@ -1,15 +1,15 @@
-import { Input } from '../View/inputViews.js';
+import { Input } from '../View/inputView.js';
 import { CarNamesValidator } from '../Validation/carNamesValidator.js';
 import { parseCarNames } from '../Model/carNamesParser.js';
 import { RacingCountValidator } from '../Validation/racingCountValidator.js';
-import { racingHandler } from '../Model/forwardProgression.js';
+import { RacingHandler } from '../Model/forwardProgression.js';
 
 export class MainController {
   constructor() {
     this.input = new Input();
   }
 
-  async startProgram() {
+  async initializeProgram() {
     const carNamesInput = await this.input.getCarNamesInput();
     new CarNamesValidator().validateCarNames(carNamesInput);
     const parsedCarNames = parseCarNames(carNamesInput);
@@ -17,6 +17,6 @@ export class MainController {
     const racingCount = await this.input.getRacingCountInput();
     new RacingCountValidator().validateRacingCount(racingCount);
 
-    new racingHandler().progressRacing(parsedCarNames, racingCount);
+    new RacingHandler().startRace(parsedCarNames, racingCount);
   }
 }
